SQL_and_HDFS/requirements.txt
autobadger @ git+https://gitlab:gldt-p2AzX_9vHWtcfx81WfDo@git.doit.wisc.edu/cdis/cs/courses/cs544/s25/tools/autobadger.git@master
grpcio
grpcio-tools
mysql-connector-python
pandas
pyarrow
requests
pymysql
sqlalchemy
retry
cryptography
